+++ packages/eslint-config-terra/eslint.config.js
@@ -67,7 +67,7 @@ module.exports = {
       },
     },
     {
-      files: ['**/wdio/**/*-spec.*'],
+      files: ['**/wdio/**/*-spec.*', '**/full-stack/**/*-spec.*'],
       rules: {
         'no-unused-expressions': 'off',
       },
@@ -75,6 +75,7 @@ module.exports = {
         after: true,
         before: true,
         browser: true,
+        FullStack: true,
         Terra: true,
         $: true,
       },