+++ packages/terra-cli/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.7.0 - (April 23, 2021)
+
+* Fixed
+  * Add back "strict" to the CLI API but only for commands. Options aren't validated, but commands are.
+
 ## 1.6.0 - (February 25, 2021)
 
 * Changed